lsp: split handlers.rs into topic-specific submodules
Converts the 1795-line crates/harn-lsp/src/handlers.rs into a
handlers/ directory with one submodule per concern. The single
`impl tower_lsp::LanguageServer for HarnLsp` block lives in
mod.rs and delegates each method to inherent `handle_*` helpers
defined in the appropriate submodule.

Split:
- completion.rs  — dot + discriminator completions
- definition.rs  — goto-definition, references, rename, cross-file resolve
- hover.rs       — hover, signature help, inlay hints
- lifecycle.rs   — initialize/initialized/shutdown + document sync
- formatting.rs  — document formatting + code-action quick-fixes
- symbols.rs     — document symbols, workspace symbols, semantic tokens
- mod.rs         — trait-impl wiring

Every file is under 800 lines. Behavior is unchanged — no logic
was rewritten, only relocated. All 21 crate tests pass and
`make all` succeeds.

Closes #57.
